Restructure ready for monorepo
In order to transition to a monorepo, I've detached the vite build from
tauri, and set up the frontend as a yarn workspace. This also involved
setting up the tsconfig in monorepo mode.

Now that this is in place, I can start moving things into separate
packages, and slowly decompose the monolith frontend into smaller parts