Migrate to CSSWG Nesting spec, reduce nesting
The nesting operator can be used for class construction in SASS/SCSS,
but in the CSSWG Nesting spec, this can't be done. The up-side of this
is that cascades are easier to identify. As for the use of imports, this
is different from component-scoped CSS.

Component-scoped CSS wouldn't get processed in the same postcss passes
as eachother, because they get compiled to CSS-in-JS. However,
CSS imports are evaluated in the same PostCSS execution pass.